// web/platform-admin-web/src/api/model/announcementResponseDtoAudience.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App API
 * The AI Base App Gateway API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 공지 대상
 */
export type AnnouncementResponseDtoAudience = typeof AnnouncementResponseDtoAudience[keyof typeof AnnouncementResponseDtoAudience];


export const AnnouncementResponseDtoAudience = {
  ALL: 'ALL',
  PLATFORM: 'PLATFORM',
  ORGANIZATION: 'ORGANIZATION',
} as const;