// web/platform-admin-web/src/api/model/announcementResponseDtoCategory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App API
 * The AI Base App Gateway API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 공지 분류
 */
export type AnnouncementResponseDtoCategory = typeof AnnouncementResponseDtoCategory[keyof typeof AnnouncementResponseDtoCategory];


export const AnnouncementResponseDtoCategory = {
  NOTICE: 'NOTICE',
  MAINTENANCE: 'MAINTENANCE',
  SECURITY: 'SECURITY',
  EVENT: 'EVENT',
} as const;