// web/platform-admin-web/src/api/model/announcementResponseDtoChannel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App API
 * The AI Base App Gateway API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 공지 채널
 */
export type AnnouncementResponseDtoChannel = typeof AnnouncementResponseDtoChannel[keyof typeof AnnouncementResponseDtoChannel];


export const AnnouncementResponseDtoChannel = {
  IN_APP: 'IN_APP',
  EMAIL: 'EMAIL',
  PUSH: 'PUSH',
} as const;