// web/platform-admin-web/src/api/model/announcementResponseDtoPriority.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App API
 * The AI Base App Gateway API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 공지 우선순위
 */
export type AnnouncementResponseDtoPriority = typeof AnnouncementResponseDtoPriority[keyof typeof AnnouncementResponseDtoPriority];


export const AnnouncementResponseDtoPriority = {
  LOW: 'LOW',
  NORMAL: 'NORMAL',
  HIGH: 'HIGH',
} as const;