// web/platform-admin-web/src/api/model/announcementResponseDtoStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App API
 * The AI Base App Gateway API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 게시 상태
 */
export type AnnouncementResponseDtoStatus = typeof AnnouncementResponseDtoStatus[keyof typeof AnnouncementResponseDtoStatus];


export const AnnouncementResponseDtoStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
} as const;